+++ uo_sys_framework/ACREModule/cfgAttributeCategories/Object.hpp
@@ -92,5 +92,15 @@ class EGVAR(ACRE,UnitOptions) {
             condition = "objectControllable";
             defaultValue = "0";
         };
+        class EGVAR(ACRE,AdditionalLanguage) {
+            displayName = "Additional Language";
+            tooltip = "Selected Language will be added to the list of spoken Babbel langauges. The language must be defined in the main acre settings.";
+            property = QEGVAR(ACRE,AdditionalLanguage);
+            control = QEGVAR(ACRE,AdditionalLanguageAttribute);
+            typeName = "STRING";
+            expression = ENTITY_EXPRESSION;
+            condition = "objectControllable";
+            defaultValue = "'None'";
+        };
     };
 };